+++ packages/common/metalk8s-sosreport/containerd.py
@@ -1,6 +1,10 @@
 #! /bin/env python3
 
-from sos.plugins import Plugin, RedHatPlugin
+# sos plugin layout changed in sos 4.0
+try:
+    from sos.report.plugins import Plugin, RedHatPlugin
+except ImportError:
+    from sos.plugins import Plugin, RedHatPlugin
 
 
 class containerd(Plugin, RedHatPlugin):